Refactored code in preparation for API split
